+++ src/utilization/epoch.ts
@@ -1,0 +1,6 @@
+/**
+ * This function returns the current timestamp in milliseconds
+ */
+export function epoch(): number {
+  return new Date().getTime();
+}